Cohort analysis on more that one page
The prevous Cohort framework could do tests only in one page because
the onInitialize event was called every time, even if a user was in a
certain Cohort in the previous page.
Now onInitialize has the parameter "new_to_this_test" so that you can
call the event "cohort + ' | Total'" only when the user enters to the
test, and not all the times the test is initialized.